% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeforecast.R
\name{detect_anormal.timesuperin}
\alias{detect_anormal.timesuperin}
\title{Detect anomaly data}
\usage{
detect_anormal.timesuperin(object, newdata, level = 0.95, value,
  direction = "both", cumul.thre = NULL)
}
\arguments{
\item{object}{result of model fitting functions.}

\item{newdata}{dataframe for predictions. Must have columns date type.}

\item{level}{Tolerance/confidence level.}

\item{value}{Anormal confirmation data.}

\item{direction}{Anomaly detection direction : 'upper', 'lower', 'both'}

\item{cumul.thre}{threshold for Cumulative residual}
}
\description{
Detection for anomaly data based on predcition interval
}
\keyword{Detection}
