% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeforecast.R
\name{pred.table.timesuperin}
\alias{pred.table.timesuperin}
\title{Model predictions}
\usage{
pred.table.timesuperin(object, newdata, level = 0.95)
}
\arguments{
\item{object}{result of model fitting functions.}

\item{newdata}{dataframe for predictions. Must have columns date type.}

\item{level}{Tolerance/confidence level.}
}
\description{
This function for predictions from the results of model fitting functions.
}
\keyword{predict}
